manual partitioning IFS bugfix
took me ages to track down this bug.

Due to IFS issues, the user choice RET was not matched up with the available
choices, so the choice key could not be returned. So, by default it would
return the RET itself, which would result in a plugin error.

The plugin error would result in ask_user returning a failed exitcode, which
in turn result in a "No root filesystem" message.