Mask live-tools.service within di-live package
This commit masks the live-tools.service by creating a symlink from
/etc/systemd/system/live-tools.service (thus completely disabling
live-tool.service).

This is required because live-tools.service partially replicates di-live
behaviour, so is redundant.

This code was being applied in common, but it's cleaner to include it in
di-live so that it's also removed when di-live pkg is removed.